requirements.txt
absl-py==1.0.0
numpy==1.21.4
Pillow==8.4.0
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing_extensions==4.0.0
